pyproject.toml
[tool.poetry]
name = "functionalnotifyer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8.6"
Pyrogram = "^1.1.13"
coloredlogs = "^15.0"
dataclasses-json = "^0.5.2"
PyYAML = "^5.4.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
